+++ window/src/connection.rs
@@ -4,13 +4,24 @@ use anyhow::Result as Fallible;
 use config::DimensionContext;
 use std::cell::RefCell;
 use std::rc::Rc;
+use std::sync::Mutex;
 
 thread_local! {
     static CONN: RefCell<Option<Rc<Connection>>> = RefCell::new(None);
 }
 
+fn nop_event_handler(_event: ApplicationEvent) {}
+
+static EVENT_HANDLER: Mutex<fn(ApplicationEvent)> = Mutex::new(nop_event_handler);
+
 pub fn shutdown() {
     CONN.with(|m| drop(m.borrow_mut().take()));
+}
+
+#[derive(Debug)]
+pub enum ApplicationEvent {
+    /// The system wants to open a command in the terminal
+    OpenCommandScript(String),
 }
 
 pub trait ConnectionOps {
@@ -22,6 +33,16 @@ pub trait ConnectionOps {
             }
         });
         res
+    }
+
+    fn set_event_handler(&self, func: fn(ApplicationEvent)) {
+        let mut handler = EVENT_HANDLER.lock().unwrap();
+        *handler = func;
+    }
+
+    fn dispatch_app_event(&self, event: ApplicationEvent) {
+        let func = EVENT_HANDLER.lock().unwrap();
+        func(event);
     }
 
     fn default_dpi(&self) -> f64 {