+++ window/src/os/macos/app.rs
@@ -1,0 +1,72 @@
+use crate::connection::ConnectionOps;
+use crate::macos::nsstring_to_str;
+use crate::{ApplicationEvent, Connection};
+use cocoa::appkit::NSApplicationTerminateReply;
+use objc::declare::ClassDecl;
+use objc::rc::StrongPtr;
+use objc::runtime::{Class, Object, Sel};
+use objc::*;
+
+const CLS_NAME: &str = "WezTermAppDelegate";
+
+extern "C" fn application_should_terminate(
+    _self: &mut Object,
+    _sel: Sel,
+    _app: *mut Object,
+) -> u64 {
+    log::debug!("application termination requested");
+    NSApplicationTerminateReply::NSTerminateLater as u64
+}
+
+extern "C" fn application_will_finish_launching(
+    _self: &mut Object,
+    _sel: Sel,
+    _notif: *mut Object,
+) {
+    log::debug!("application_will_finish_launching");
+}
+
+extern "C" fn application_open_file(
+    _self: &mut Object,
+    _sel: Sel,
+    _app: *mut Object,
+    file_name: *mut Object,
+) {
+    let file_name = unsafe { nsstring_to_str(file_name) }.to_string();
+    if let Some(conn) = Connection::get() {
+        conn.dispatch_app_event(ApplicationEvent::OpenCommandScript(file_name));
+    }
+}
+
+fn get_class() -> &'static Class {
+    Class::get(CLS_NAME).unwrap_or_else(|| {
+        let mut cls = ClassDecl::new(CLS_NAME, class!(NSWindow))
+            .expect("Unable to register application class");
+
+        unsafe {
+            cls.add_method(
+                sel!(applicationShouldTerminate:),
+                application_should_terminate as extern "C" fn(&mut Object, Sel, *mut Object) -> u64,
+            );
+            cls.add_method(
+                sel!(applicationWillFinishLaunching:),
+                application_will_finish_launching as extern "C" fn(&mut Object, Sel, *mut Object),
+            );
+            cls.add_method(
+                sel!(application:openFile:),
+                application_open_file as extern "C" fn(&mut Object, Sel, *mut Object, *mut Object),
+            );
+        }
+
+        cls.register()
+    })
+}
+
+pub fn create_app_delegate() -> StrongPtr {
+    let cls = get_class();
+    unsafe {
+        let delegate: *mut Object = msg_send![cls, alloc];
+        let delegate: *mut Object = msg_send![delegate, init];
+        StrongPtr::new(delegate)
+    }
+}